[FIX] fix a crash in graphview when user clicked rapidly twice on menu item
The issue was that the get_search_fields method tried to get the view's 
corresponding searchview after a rpc (fields_get).  Sadly, if the user
had meanwhile clicked again on the menu item, the first view would be
detached from the view manager and would be unable to reach the search 
view.  The solution is to move the lookup for the search view in the start
method, where it is guaranteed to exist.