[FIX] warning: make blocking warning actually doing something (opw 607380)
Instead of raising an exception (which only interupt the call of the onchange methods), return a dictionnary to set the forbidden partner or product to an empty value.
The call is not blocking in a way it forbid to create a record that uses this partner/product but it make it not possible to select it through the web interface.